requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2